aggiunta gestione admin

+++ FlavourFitAPI/src/main/java/com/example/demo/controller/InteractionController.java
@@ -1,10 +1,15 @@
 package com.example.demo.controller;
 
+import com.example.demo.model.ErrorMessage;
 import com.example.demo.model.document.Interaction;
 import com.example.demo.model.interactionCrud.CreateInteractionInput;
 import com.example.demo.service.InteractionService;
+import com.example.demo.utils.AuthorizationUtil;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,6 +21,11 @@ public class InteractionController {
 
     @Autowired
     private InteractionService service;
+    private final AuthorizationUtil authorizationUtil;
+
+    public InteractionController(AuthorizationUtil authorizationUtil) {
+        this.authorizationUtil = authorizationUtil;
+    }
 
     @GetMapping
     @SecurityRequirement(name = "bearerAuth")
@@ -43,7 +53,18 @@ public class InteractionController {
 
     @DeleteMapping("/{id}")
     @SecurityRequirement(name = "bearerAuth")
-    public void deleteInteraction(@PathVariable String id) throws Exception {
+    public ResponseEntity<ErrorMessage> deleteInteraction(@PathVariable String id, HttpServletRequest request) throws Exception {
+
+        Optional<Interaction> interaction = service.getInteractionById(id);if (interaction.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ErrorMessage("Recipe not found"));
+        }
+
+        //Recipe recipe = recipeOpt.get();
+        boolean authorized = authorizationUtil.verifyOwnershipOrAdmin(request, id);
+        if(!authorized)
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ErrorMessage("No access to delete user"));
         service.deleteInteraction(id);
+        return null;
     }
 }
